refactor: move llm/model/mock.ts to test scope
